Create JavaReadFileAndDraw
One text file contains data for rectangles with sides parallel to the coordinate axes. The first line contains a number indicating the number of figures, and the following lines contain the data for them. One line contains the X and Y data for the upper left point , followed by X and Y for the right-most point. A space is used as the delimiter. To compile a Java Application to read this file and to graphically represent the rectangles.